# projects/phoenix-api/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Phoenix Platform API
  description: |
    Phoenix Platform API for managing observability cost optimization through
    experiments, pipeline deployments, and metrics analysis.
  version: 2.0.0
  contact:
    name: Phoenix Team
    email: phoenix@example.com
servers:
  - url: http://localhost:8080/api/v1
    description: Local development
  - url: https://phoenix.example.com/api/v1
    description: Production
security:
  - bearerAuth: []
paths:
...
  /experiments:
    get:
      summary: List experiments
      operationId: listExperiments
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [initializing, running, completed, failed]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: page_size
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: List of experiments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Experiment'

    post:
      summary: Create experiment
      operationId: createExperiment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateExperimentRequest'
      responses:
        '201':
          description: Experiment created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Experiment'
  /experiments/{id}:
    get:
      summary: Get experiment details
      operationId: getExperiment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Experiment details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Experiment'
        '404':
          description: Experiment not found
  /experiments/{id}/phase:
    put:
      summary: Update experiment phase
      operationId: updateExperimentPhase
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - phase
              properties:
                phase:
                  type: string
                  enum: [initializing, running, analyzing, completed, failed]
      responses:
        '200':
          description: Phase updated
  /experiments/{id}/start:
    post:
      summary: Start experiment
      operationId: startExperiment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Experiment started
  /experiments/{id}/stop:
    post:
      summary: Stop experiment
      operationId: stopExperiment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Experiment stopped

  # Pipeline deployment endpoints
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Experiment:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        phase:
          type: string
          enum: [initializing, running, analyzing, completed, failed]
        status:
          type: string
          deprecated: true
          description: Deprecated, use phase instead
        baseline_pipeline:
          type: string
        candidate_pipeline:
          type: string
        target_nodes:
          type: array
          items:
            type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreateExperimentRequest:
      type: object
      required:
        - name
        - baseline_template
        - candidate_template
      properties:
        name:
          type: string
        description:
          type: string
        baseline_template:
          type: string
        candidate_template:
          type: string
        target_hosts:
          type: array
          items:
            type: string
        duration:
          type: string
          example: "1h"
        parameters:
          type: object
          additionalProperties: true
...